requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
google-cloud-speech>=2.20.0
google-cloud-language>=2.10.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
pytest>=7.4.0
pytest-flask>=1.2.0